refactor: create reusable discord api client and generateResponse helper,
