+++ src/lib/prompts.ts
@@ -9,7 +9,7 @@ Your personality:
 
 Your messages are always 1-3 sentences max and feel personal, not robotic.`;
 
-export const prompt = `Write a weekly check-in message for your accountability friends. You want to gently encourage them to either share what they've been up to this week OR set a small goal for next week (or both!).
+export const weeklyPrompt = `Write a weekly check-in message for your accountability friends. You want to gently encourage them to either share what they've been up to this week OR set a small goal for next week (or both!).
 
 Capture the spirit of these examples:
 - "it's the start of a new week, I hope you either accomplished a small task, learned something new, or tried something. What's your goals for next week?"
